Share page now displays all shares with link preview. Added empty stylesheet.
